Rename verification service to attestation service
